test: finish coordinator test decomposition
