+++ tests/test_coordinator.py
@@ -53,7 +53,6 @@ HOLDING_REGISTERS = {r.name: r.address for r in get_registers_by_function("03")}
 # ✅ FIXED: Import correct coordinator class name
 from custom_components.thessla_green_modbus.coordinator import (
     ThesslaGreenModbusCoordinator,
-    _PermanentModbusError,
 )
 from custom_components.thessla_green_modbus.coordinator import (
     dt_util as coordinator_dt_util,
@@ -946,133 +945,6 @@ async def test_async_setup_invalid_capabilities(coordinator):
 
     assert str(err.value) == "invalid_capabilities"
     scanner_instance.close.assert_awaited_once()
-
-
-@pytest.mark.asyncio
-async def test_coordinator_tracks_offline_and_recovers(monkeypatch) -> None:
-    """Transient failures increment counters and successful reads reset them."""
-
-    coordinator = ThesslaGreenModbusCoordinator.from_params(
-        MagicMock(),
-        "host",
-        502,
-        1,
-        "name",
-        scan_interval=5,
-        retry=1,
-    )
-    coordinator.client = MagicMock(connected=True)
-    coordinator._disconnect = AsyncMock()
-    coordinator._ensure_connection = AsyncMock()
-    coordinator._read_input_registers_optimized = AsyncMock(
-        side_effect=[ConnectionException("fail"), {"reg": 1}]
-    )
-    coordinator._read_holding_registers_optimized = AsyncMock(return_value={})
-    coordinator._read_coil_registers_optimized = AsyncMock(return_value={})
-    coordinator._read_discrete_inputs_optimized = AsyncMock(return_value={})
-
-    with pytest.raises(UpdateFailed):
-        await coordinator._async_update_data()
-
-    assert coordinator._consecutive_failures == 1  # nosec: explicit state check
-    coordinator._read_input_registers_optimized.reset_mock(side_effect=True)
-    coordinator._read_input_registers_optimized.side_effect = None
-    coordinator._read_input_registers_optimized.return_value = {"reg": 1}
-
-    data = await coordinator._async_update_data()
-
-    assert data["reg"] == 1  # nosec: explicit state check
-    assert coordinator._consecutive_failures == 0  # nosec: explicit state check
-    assert coordinator.statistics["last_successful_update"] is not None  # nosec
-
-
-@pytest.mark.asyncio
-async def test_coordinator_disconnects_after_retries(monkeypatch) -> None:
-    """Persistent failures force a disconnect and surface an error."""
-
-    coordinator = ThesslaGreenModbusCoordinator.from_params(
-        MagicMock(),
-        "host",
-        502,
-        1,
-        "name",
-        scan_interval=5,
-        retry=1,
-    )
-    coordinator._max_failures = 1
-    coordinator.client = MagicMock(connected=True)
-    coordinator._disconnect = AsyncMock()
-    coordinator._ensure_connection = AsyncMock()
-    coordinator._read_input_registers_optimized = AsyncMock(side_effect=TimeoutError("boom"))
-    coordinator._read_holding_registers_optimized = AsyncMock(return_value={})
-    coordinator._read_coil_registers_optimized = AsyncMock(return_value={})
-    coordinator._read_discrete_inputs_optimized = AsyncMock(return_value={})
-
-    with pytest.raises(UpdateFailed):
-        await coordinator._async_update_data()
-
-    coordinator._disconnect.assert_awaited_once()
-    assert coordinator.client.connected  # nosec: explicit state check
-
-
-@pytest.mark.asyncio
-async def test_read_with_retry_retries_transient_errors():
-    """Coordinator retries transient read errors before succeeding."""
-
-    coordinator = ThesslaGreenModbusCoordinator.from_params(
-        MagicMock(),
-        "host",
-        502,
-        1,
-        "name",
-        retry=2,
-    )
-    coordinator._disconnect = AsyncMock()
-    response = MagicMock()
-    response.isError.return_value = False
-    coordinator._call_modbus = AsyncMock(side_effect=[TimeoutError("boom"), response])
-
-    result = await coordinator._read_with_retry(
-        lambda *_args, **_kwargs: None,  # pragma: no cover - not invoked directly
-        10,
-        1,
-        register_type="input",
-    )
-
-    assert result is response  # nosec: explicit state check
-    assert coordinator._call_modbus.await_count == 2
-    # _disconnect is not called on TimeoutError without an active transport (production behaviour)
-
-
-@pytest.mark.asyncio
-async def test_read_with_retry_skips_illegal_data_address():
-    """Illegal data address errors should not be retried."""
-
-    coordinator = ThesslaGreenModbusCoordinator.from_params(
-        MagicMock(),
-        "host",
-        502,
-        1,
-        "name",
-        retry=3,
-    )
-    coordinator._disconnect = AsyncMock()
-
-    response = MagicMock()
-    response.isError.return_value = True
-    response.exception_code = 2
-    coordinator._call_modbus = AsyncMock(return_value=response)
-
-    with pytest.raises(_PermanentModbusError):
-        await coordinator._read_with_retry(
-            lambda *_args, **_kwargs: None,  # pragma: no cover - not invoked directly
-            10,
-            1,
-            register_type="input",
-        )
-
-    assert coordinator._call_modbus.await_count == 1
-    coordinator._disconnect.assert_not_awaited()
 
 
 class TestParseBackoffJitter:
